[feat] Adding a Launcher class for main method
